requirements.txt
fastapi
uvicorn[standard]
pytest
pydantic
httpx
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]